requirements.txt
faiss==1.7.1
numpy==1.23.1
pandas==1.5.0
pytest==7.1.3
pywebio==1.7.0
scikit_learn==1.1.3
scikit_surprise==1.1.3
